update connection management for connection pool
If the connection belongs to a connection pool, it must be
returned to the pool when calling "close" without actually
closing the connection. All connections will be closed when
poll:close() will be called. The same behavior is used in
JDBC connection pool.
With this change, an attempt to "put" an unusable connection
is no longer valid.

Code that doesn't test anything has been removed from the tests.

Fixed #33