add a check for the result of updating the connection authentication settings
We don't check the result of updating the connection authentication
settings in the conn: reset() method and failure will be detected
only when trying to execute a request. But in programming, it is a
good practice to check the result of the function.
The patch adds a check on the result of updating the connection
authentication settings. On failure, an exception will be thrown.