make lock arithmetic proper
Make lock arithmetic proper and complite in conn:execute(),
conn:close(), conn:reset() and conn:quote(). Release lock in
pool:put(conn). The absence of this operation could lead to an
infinite bloсking when used with fiber as describerd in the case
below.

A connection conn is acquired from a pool and we acquire a lock.
Then we start the first fiber with pool:put(conn)`, which try to
acquire lock and yield. Then the second fiber is started to
perform conn:execute(), conn:reset() and conn:quote() which try
to acquire lock, because the connection is still marked as
`usable` and yield too. Before this patch, it's appeared that the
lock is never released.

Closes #34